스프링시큐리티 인가설정

+++ src/main/java/com/example/whosbookupdate/dto/MemberResponseDto.java
@@ -8,7 +8,7 @@ import java.util.Date;
 @Data
 public class MemberResponseDto {
 
-    private int memberId;
+    private Long memberId;
     private String email;
     private String nickname;
     private String password;
